Auth Helper - Authorize returning 401
Modified signup route to auto log in on submitting
new user form. This works...

Also added button to 'profile' function in UserController
This ideally hits the user show route (/user/:id) and that
functionality works. The middleware to check if currentuser
is authorized to view chosen page does not work - it instead
returns 401 every time, likely because current user isn't
able to be passed to authorize route.

Currently working on passing current user (json attached to
the get request?) so comparison can be made to check if valid
user.